Big refactor
Refactored the project so that the chip8 core doesnt depend on any
dependencies and added Emulator struct which actually uses the chip8
state to render the window, handle input and audio (not yet audio).

I might extract the chip8 core to a library, while the emulator binary
will use winit + something in the future.

This is sort of WIP commit, a lot unopimized and uncool things I left
out but will fix in the next one.